+++ Core/Runtime/Scripts/DetectionUtilities/PalmDirectionDetector.cs
@@ -167,12 +167,12 @@ namespace Leap.Unity
             switch (PointingType)
             {
                 case PointingType.RelativeToHorizon:
-                    Quaternion cameraRot = Camera.main.transform.rotation;
+                    Quaternion cameraRot = MainCameraProvider.Instance.transform.rotation;
                     float cameraYaw = cameraRot.eulerAngles.y;
                     Quaternion rotator = Quaternion.AngleAxis(cameraYaw, Vector3.up);
                     return rotator * PointingDirection;
                 case PointingType.RelativeToCamera:
-                    return Camera.main.transform.TransformDirection(PointingDirection);
+                    return MainCameraProvider.Instance.transform.TransformDirection(PointingDirection);
                 case PointingType.RelativeToWorld:
                     return PointingDirection;
                 case PointingType.AtTarget: